refactor reorder logic of dragAndDrop and move functions to model
