+++ app/assets/javascripts/models/beta/projectBoard.js
@@ -2,7 +2,7 @@ import httpService from '../../services/httpService';
 import changeCase from 'change-object-case';
 import * as Story from './story';
 import * as Project from './project';
-import { operands } from './../../libs/beta/constants';
+import { operands, status } from './../../libs/beta/constants';
 import { values } from 'underscore';
 
 export const get = async (projectId) => {
@@ -51,7 +51,7 @@ export const translateWord = (operand, word, translations) =>
     ? word
     : I18n.t(`story.${operand}.${translations[word] || word}`, { locale: I18n.defaultLocale });
 
-const translatedOperands = ['type','state'];
+const translatedOperands = ['type', 'state'];
 
 export const haveTranslation = operand => translatedOperands.includes(operand)
 
@@ -64,3 +64,34 @@ export const toggleColumn = (projectBoard, column, callback) => {
     return callback.onToggle();
   }
 }
+
+// Drag And drop utils
+const calculatePosition = (aboveStory, bellowStory) => {
+  if (bellowStory === undefined) return (Number(aboveStory.position) + 1);
+  if (aboveStory === undefined) return (Number(bellowStory.position) - 1);
+  return (Number(bellowStory.position) + Number(aboveStory.position)) / 2;
+}
+
+export const getNewPosition = (destinatitonIndex, sourceIndex, storiesArray, isSameColumn, storyType) => {
+  //TODO: remove this second condition later
+  if (!isSameColumn && storyType !== 'feature') {
+    return calculatePosition(storiesArray[destinatitonIndex - 1], storiesArray[destinatitonIndex]);
+  }
+  if (sourceIndex > destinatitonIndex) {
+    return calculatePosition(storiesArray[destinatitonIndex - 1], storiesArray[destinatitonIndex]);
+  }
+  return calculatePosition(storiesArray[destinatitonIndex], storiesArray[destinatitonIndex + 1]);
+}
+
+// reorder the array
+export const moveTask = (sourceArray, destinationArray, sourceIndex, destinationIndex) => {
+  const newSourceArray = sourceArray;
+  const [removed] = newSourceArray.splice(sourceIndex, 1);
+  const newDestinationArray = destinationArray;
+  newDestinationArray.splice(destinationIndex, 0, removed);
+  return [...newDestinationArray];
+}
+
+export const getNewSprints = (newStories, sprints) => sprints.map((sprint, index) => index === 0 ? { ...sprint, stories: newStories } : sprint);
+
+export const getNewState = column => column === 'chillyBin' ? status.UNSCHEDULED : status.UNSTARTED;
